Fix warnings

+++ common_helper/common_helper.cpp
@@ -55,12 +55,12 @@ template<typename T>
 T& CommonHelper::GetValue(std::vector<T>& val_list, std::vector<int32_t> shape, std::vector<int32_t> pos)
 {
     static std::vector<int32_t> stride(shape.size());
-    for (int32_t i = 0; i < stride.size(); i++) {
+    for (size_t i = 0; i < stride.size(); i++) {
         stride[i] = std::accumulate(shape.begin() + i + 1, shape.end(), 1, std::multiplies<int32_t>());
     }
 
     int32_t index = 0;
-    for (int32_t i = 0; i < pos.size(); i++) {
+    for (size_t i = 0; i < pos.size(); i++) {
         index += pos[i] * stride[i];
     }
     return val_list.at(index);
@@ -74,20 +74,20 @@ template uint8_t& CommonHelper::GetValue<uint8_t>(std::vector<uint8_t>& val_list
 template<typename T>
 void CommonHelper::PrintValue(std::vector<T>& val_list, std::vector<int32_t> shape)
 {
-    int32_t element_num = std::accumulate(shape.begin(), shape.end(), 1, std::multiplies<int32_t>());
+    size_t element_num = std::accumulate(shape.begin(), shape.end(), 1, std::multiplies<int32_t>());
     if (val_list.size() != element_num) {
-        PRINT_E("[PrintValue] invalid data size. %zd, %d\n", val_list.size(), element_num);
+        PRINT_E("[PrintValue] invalid data size. %zd, %zd\n", val_list.size(), element_num);
         return;
     }
 
     std::vector<int32_t> stride(shape.size() - 1);
-    for (int32_t i = 0; i < stride.size(); i++) {
+    for (size_t i = 0; i < stride.size(); i++) {
         stride[i] = std::accumulate(shape.begin() + i + 1, shape.end(), 1, std::multiplies<int32_t>());
     }
 
-    for (int32_t i = 0; i < element_num; i++) {
+    for (size_t i = 0; i < element_num; i++) {
         printf("%3.4f, ", (float)val_list.at(i));
-        for (int32_t j = 0; j < stride.size(); j++) {
+        for (size_t j = 0; j < stride.size(); j++) {
             if ((i + 1) % stride[j] == 0) printf("\n");
         }
     }
